prepperapp/content/scripts/requirements.txt
# PrepperApp Content Processing Requirements

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF Processing
pdfplumber>=0.11.4
pypdf>=3.17.0
pdf2image>=1.17.0
camelot-py[cv]>=0.11.0  # For complex table extraction

# Image Processing
python-magic>=0.4.27
Pillow>=10.0.0
pillow-heif>=0.19.0
PyMuPDF>=1.26.0  # For PDF image extraction without poppler

# OCR (optional, for scanned PDFs)
pytesseract>=0.3.13

# Text Processing
scikit-learn>=1.5.2  # For TF-IDF keyword extraction
nltk>=3.9.1

# ZIM file handling
libzim>=3.0.0

# Compression
zstandard>=0.22.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Progress bars
tqdm>=4.66.0

# JSON processing
jsonlines>=4.0.0

# Performance profiling (optional)
memory-profiler>=0.61.0